Centralize ZK Types
