+++ src/zk/methods/getCompressedAccount.ts
@@ -1,12 +1,9 @@
 import type { RpcCaller } from "../../rpc/caller";
 import type {
+  GetCompressedAccountFn,
   GetCompressedAccountRequest,
   GetCompressedAccountResponse,
 } from "../types";
-
-export type GetCompressedAccountFn = (
-  p: GetCompressedAccountRequest
-) => Promise<GetCompressedAccountResponse>;
 
 export const makeGetCompressedAccount =
   (call: RpcCaller): GetCompressedAccountFn =>
