+++ src/zk/methods/getCompressedAccountsByOwner.ts
@@ -1,12 +1,11 @@
 import { RpcCaller } from "../../rpc/caller";
 import {
+  GetCompressedAccountsByOwnerFn,
   GetCompressedAccountsByOwnerRequest as Req,
   GetCompressedAccountsByOwnerResponse as Res,
 } from "../types";
 
-export type GetCompressedAccountsByOwnerFn = (p: Req) => Promise<Res>;
-
 export const makeGetCompressedAccountsByOwner =
   (call: RpcCaller): GetCompressedAccountsByOwnerFn =>
-  (params) =>
+  (params: Req) =>
     call<Req, Res>("getCompressedAccountsByOwner", params);
