+++ src/zk/methods/getMultipleNewAddressProofsV2.ts
@@ -1,12 +1,9 @@
 import type { RpcCaller } from "../../rpc/caller";
 import type {
+  GetMultipleNewAddressProofsV2Fn,
   GetMultipleNewAddressProofsV2Request,
   GetMultipleNewAddressProofsV2Response,
 } from "../types";
-
-export type GetMultipleNewAddressProofsV2Fn = (
-  p: GetMultipleNewAddressProofsV2Request
-) => Promise<GetMultipleNewAddressProofsV2Response>;
 
 export const makeGetMultipleNewAddressProofsV2 =
   (call: RpcCaller): GetMultipleNewAddressProofsV2Fn =>
